requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
async-timeout==4.0.3
asyncpg==0.29.0
black==23.11.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
exceptiongroup==1.1.3
Faker==20.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iso8601==1.1.0
jedi==0.19.1
Jinja2==3.1.2
litestar==2.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgspec==0.18.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
platformdirs==4.0.0
polyfactory==2.12.0
prompt-toolkit==3.0.41
ptpython==3.0.23
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.7.0
rich-click==1.7.1
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
wcwidth==0.2.10
